install: validate the git tree with GPG before installing anything
This lessens the trust requirements on the service hosting the git
repository (personal server or GitHub) and about transmission protocols
(like unsecure git://).  This is quite needed as I don't want a
potential attacker to be able to change my home files without me
noticing.

Also, put my public key in this git repository so that an Internet
connection is not needed to perform the first install.