*: add copyright information to almost every script
License the scripts I wrote with the MIT license so that other people
can use them without worrying much about the legal aspect.

I did not put a license on really short scripts (less than 10 lines of
code) as this just does not make any sense.  It is not intellectual
property to only read the manual and add a one-liner script in $PATH to
ease the use of FOSS tools (Free and Open Source Software).  Anyway if
this cause trouble, feel free to contact me if you want me to put a
clear free license on these files.

A good side-effect of the copyright information is that it allows easy
tracking of old scripts and the ones which are more recent (through the
copyright year).