bin/makecleanpkg: move fake sudo into a real program
Diverting calls to sudo into a shell function appears to be quite buggy,
as the fakeroot environment gets broken, "ps" shows a non-intuituve
process hierarchy and there were issues when using namcap to check the
compiled packages.

Move the function into a shell script and use $PATH to divert the use of
sudo to this script.

While at it, hook pacman too in order to always run it with a fake
chroot() syscall.