bin/tpm-show: describe more TPM 2.0 response codes
On https://gitter.im/tpm2-software/community someone reported a TPM
issue with:

    WARNING:esys:src/tss2-esys/api/Esys_Sign.c:311:Esys_Sign_Finish() Received TPM Error
    ERROR:esys:src/tss2-esys/api/Esys_Sign.c:105:Esys_Sign() Esys Finish ErrorCode (0x000001d5)
    ERROR: Esys_Sign: tpm:parameter(1):structure is the wrong size

Being able to know that 0x1d5 is the sum of:

* TPM_RC_1 = 0x100
* TPM_RC_FMT1 = 0x080
* TPM_RC_P = 0x040
* TPM_RC_SIZE=TPM_RC_FMT1 + 0x015

is not straightforward. Expand the response code to include for example:

    TPM2_RC_PARAM1_SIZE = 0x1d5

which would make understanding rc=0x1d5 easier.

While at it, decode the previosly-unknown response codes:

* GetCapability->0x2cb is TPM_RC_P + TPM_RC_2 + TPM_RC_HANDLE (bad handle in param 2)
* Public_Read->0x184 is TP2_RC_H + TPM_RC_1 + TPM_RC_VALUE (bad value in handle 1)
* NV_Read->0x9a2 is TPM_RC_S + TPM_RC_1 + TPM_RC_BAD_AUTH (bad auth in session 1)