bin/nsenter-podman-net: enter UTS namespace too
The hostname is used in some network protocols (such as DHCP). It also
helps identifying the container ID as the shell prompt gets modified
with the hostname.

Use `nsenter --uts` to enter the UTS namespace of the container at the
same time as its network namespace.

While at it, document the configuration for `sudo`.